Dev product

+++ listacheck/src/main/java/com/nocountryc14/listacheck/controller/ProductController.java
@@ -1,6 +1,8 @@
 package com.nocountryc14.listacheck.controller;
 
-import com.nocountryc14.listacheck.model.Product;
+
+import com.nocountryc14.listacheck.dto.ProductDto;
+
 import com.nocountryc14.listacheck.service.IProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -8,42 +10,76 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
+@RequestMapping("/product")
 public class ProductController {
 
     @Autowired
     private IProductService prodService;
 
 
-    //CREAR
-    @PostMapping("/products/create")
-    public String createProduct(@RequestBody Product prod) {
-        prodService.saveProduct(prod);
-        //devuelvo String q se creo correctamente
-        return "El producto fue creado correctamente";
+    //create
+    @PostMapping("/create")
+    public String createProduct(@RequestBody ProductDto prodDto) {
+        prodService.createProduct(prodDto);
+
+        return "The product has been created successfully.";
 
     }
 
-    //consulta-read
-   @GetMapping("/products/get")
-    public List<Product> getProduct() {
+    //get
+    @GetMapping("/get")
+    public String getProduct() {
 
-        return prodService.getProduct();
+        List<ProductDto> productDto = prodService.getProducts();
+
+        if (productDto.isEmpty()) {
+            return "No product has been added yet.";
+        } else {
+            return productDto.toString();
+        }
     }
 
     //DELETE/BAJA
-    @DeleteMapping("/product/delete/{id}")
+    @DeleteMapping("/delete/{id_product}")
     public String deleteProduct(@PathVariable Long id_product) {
-        prodService.deleteProduct(id_product);
-        //devuelvo String q se creo correctamente
-        return "El producto fue borrado correctamente";
+        ProductDto productDto = prodService.findProductById(id_product);
 
+        if (productDto != null) {
+            prodService.deleteProduct(id_product);
+            return "The product has been deleted successfully";
+        } else {
+            return "There is no product with this ID to delete.";
+        }
 
 
     }
 
-    //BUSQUEDA por id
-    @GetMapping("/product/{id}")
-    public void findProduct(Long id_product) {
-               prodService.findProduct(id_product);
+    // Find by ID
+    @GetMapping("/{id_product}")
+    public String findProduct(@PathVariable Long id_product) {
+        ProductDto productDto = prodService.findProductById(id_product);
+
+        if (productDto != null) {
+            return productDto.toString();
+
+        } else {
+            return "There is no product with this ID.";
+        }
     }
+
+    // Update
+    @PutMapping("/update/{id_product}")
+    public String updateProduct(@PathVariable Long id_product, @RequestBody ProductDto updatedProductDto) {
+
+        ProductDto productDto = prodService.findProductById(id_product);
+
+        if (productDto != null) {
+            prodService.updateProduct(id_product, updatedProductDto);
+            return "The product has been updated successfully.";
+        } else {
+            return "There is no product with this ID to update.";
+        }
+    }
+
+
 }
